Clean up theme, add overrides for control inputs
